+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## UNRELEASED
+## 1.9.15 - 2026-03-25
 - Images loaded with Interplanar Projector are now limited to maximum of 10000x10000.
+- While playing on the multiplayer server - projecting images from URL will only work for allowed domains by default (configurable). This is a client-only setting, so players can change it at any time.
+- Added a config option to allow projecting images from file only relatively to game directory (while on multiplayer server).
 
 ## 1.9.14 - 2026-03-04
 - Added compatibility with **Easy Anvils** for Interplanar Projector extended renaming length.